+++ services/src/main/java/config/MainConfig.java
@@ -1,10 +1,8 @@
 package config;
 
-import api.TireFacade;
-import fascadeImpl.TireFacadeImpl;
+import facade.TireFacadeImpl;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import services.MappingService;
 
 /***
  * @author Jakub Palenik, 422453@mail.muni.cz
